# broker/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Autolinks Broker Service

## the domain of the service
#host: 127.0.0.1:10010
#
## array of all schemes that the API supports
#schemes:
#  # - https
#  - http
#
## will be prefixed to all paths (comment for no prefix)
#basePath: /

# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

# define complex objects which can be references with $ref: "#/definitions/X"
definitions:
...
  Service:
    type: object
    properties:
      name:
        type: string
        description: the name of the service
      location:
        type: string
        description: where to find the service
      description:
        type: string
        description: summray fo the service
      endpoints:
        type: array
        description: the endpoints a service offers
        items:
          $ref: '#/definitions/Endpoint'
  Endpoint:
    type: object
    properties:
      name:
        type: string
        description: the name of the endpoint
      requireslogin:
        type: boolean
        description: does the service require a user to be logged in?

  # objects related to user management
...
  Error:
    type: object
    properties:
      message:
        type: string
      fields:
        type: string
securityDefinitions:
#  broker_auth:
#    type: "oauth2"
#    authorizationUrl: "/oauth/dialog"
#    flow: "implicit"
#    scopes:
#      write:data: "modify data in your account"
#      read:data: "read your data"
  # api_key:
  #   type: "apiKey"
  #   name: "api_key"
  #   in: "header"
  broker_auth:
    type: "basic"
...
paths:
  /service/registerService:
    # tell nodejs where to find the code for this path
    x-swagger-router-controller: services
    post:
      operationId: register_service
      tags:
        - Service
      description: Register a service
      parameters:
        - name: service
          in: body
          description: The service to be added
          schema:
            $ref: "#/definitions/Service"
      responses:
        200:
          description: Success
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...